feat: redirect to new portfolio
